All test types now use the 'reportSubscriptionToLocust' function
