+++ ably/personal_task.go
@@ -2,14 +2,12 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"log"
 	"math/rand"
 	"strconv"
 	"time"
 
 	"github.com/ably-forks/boomer"
-	"github.com/ably/ably-go/ably"
 )
 
 var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
@@ -33,42 +31,6 @@ func randomDelay() {
 	time.Sleep(time.Duration(r) * time.Second)
 }
 
-func reportSubscriptionToLocust(ctx context.Context, sub *ably.Subscription, conn *ably.Conn) {
-	defer sub.Close()
-
-	connectionStateChannel := make(chan ably.State)
-	conn.On(connectionStateChannel)
-
-	var lastDisconnectTime int64 = 0
-
-	for {
-		select {
-		case connState := <-connectionStateChannel:
-			if connState.State == ably.StateConnDisconnected {
-				lastDisconnectTime = millisecondTimestamp()
-			} else if connState.State == ably.StateConnConnected && lastDisconnectTime != 0 {
-				timeDisconnected := millisecondTimestamp() - lastDisconnectTime
-
-				boomer.RecordSuccess("ably", "reconnect", timeDisconnected, 0)
-			}
-		case <-ctx.Done():
-			return
-		case msg := <-sub.MessageChannel():
-			timePublished, err := strconv.ParseInt(msg.Name, 10, 64)
-
-			if err != nil {
-				boomer.RecordFailure("ably", "subscribe", 0, err.Error())
-				break
-			}
-
-			timeElapsed := millisecondTimestamp() - timePublished
-			bytes := len(fmt.Sprint(msg.Data))
-
-			boomer.RecordSuccess("ably", "subscribe", timeElapsed, int64(bytes))
-		}
-	}
-}
-
 func personalTask(testConfig TestConfig) {
 	ctx, cancel := context.WithCancel(context.Background())
 	boomer.Events.Subscribe("boomer:stop", cancel)
@@ -88,6 +50,7 @@ func personalTask(testConfig TestConfig) {
 			defer subClient.Close()
 
 			channel := subClient.Channels.Get(channelName)
+			defer channel.Close()
 
 			sub, err := channel.Subscribe()
 			if err != nil {
@@ -107,6 +70,7 @@ func personalTask(testConfig TestConfig) {
 	defer publishClient.Close()
 
 	channel := publishClient.Channels.Get(channelName)
+	defer channel.Close()
 
 	randomDelay()
